currency_convertor/requirements.txt
Django
djangorestframework
adrf
psycopg2-binary
httpx
django-environ
factory-boy
Faker